Don't use a shell to run a command as a list
Example: this hangs forever:

    raven-cron -- python -c "omg an error!"

because only "python" is passed to the shell, so it will start listening
from stdin and never finish. In general, arguments are stripped.
Wrapping everything in quotes works ok, and it's the only way to support
pipes etc I guess:

    raven-cron 'python -c "omg an error!"'

but the first option is useful too because the second requires extra
escaping etc.